--- client/src/pages/Auth/Login.css
@@ -1,34 +1,0 @@
-/* .login {
-  display: grid;
-  grid-template-columns: repeat(15, 1fr);
-  background-size: cover;
-  background-repeat: no-repeat;
-  grid-column: 1 / span-1;
-  background-position: bottom center;
-  grid-column-gap: 20px;
-  height: 100vh;
-  width: 100%;
-  max-width: 1206;
-  margin-right: auto;
-  margin-left: auto;
-}
-.input-group {
-  text-align: center;
-  margin-top: 300px;
-  grid-column: 6/11;
-}
-input {
-  height: 36px;
-  width: 80%;
-  background: none;
-  border: 0.5px solid black;
-  margin-top: -40px;
-  text-align: center;
-}
-button {
-  margin-top: 20px;
-}
-
-input:focus {
-  outline: none;
-} */
