+++ client/src/pages/Auth/Login/Login.css
@@ -1,0 +1,32 @@
+.FeatureContainer_image.Login {
+  background-image: url('../../../assets/images/hairwhite.jpg');
+}
+
+.Login_inputGroup {
+  text-align: center;
+  margin-top: 245px;
+  grid-column: 6/10;
+  display: flex;
+  flex-direction: column;
+}
+.Login_form h5 {
+  margin: 20px;
+  font-size: 1.4rem;
+  line-height: normal;
+}
+.Login_input {
+  height: 32px;
+  width: 100%;
+  background: none;
+  border: 0.5px solid black;
+  margin-top: -40px;
+  text-align: center;
+  margin-bottom: -5px;
+}
+.Login.btn--btn-primary {
+  margin-top: 30px;
+  padding-top: 0;
+}
+.Login p {
+  margin-top: 20px;
+}
