+++ client/src/pages/Auth/UserRegistration/UserRegistration.css
@@ -1,5 +1,5 @@
 .FeatureContainer_image.Profile {
-  background-image: url('../../assets/images/hairgrey90.jpg');
+  background-image: url('../../../assets/images/hairgrey90.jpg');
   margin-bottom: -10px;
 }
