+++ client/src/pages/Auth/UserRegistration/UserRegistration.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import UserImage from '../../UIElements/UserImage'
+import UserImage from '../../../UIElements/UserImage'
 import VisibilityRoundedIcon from '@material-ui/icons/VisibilityRounded'
 import VisibilityOffRoundedIcon from '@material-ui/icons/VisibilityOffRounded'
 import Input from '@material-ui/core/Input'
@@ -9,10 +9,10 @@ import { toast } from 'react-toastify'
 import { NavLink } from 'react-router-dom'
 import { useDispatch } from 'react-redux'
 import { useHistory } from 'react-router-dom'
-import { AUTH } from '../../constants/actionTypes'
-import * as api from '../../api/index'
+import { AUTH } from '../../../constants/actionTypes'
+import * as api from '../../../api/index'
 import './UserRegistration.css'
-import '../Profile/Profile.css'
+import '../../Profile/Profile.css'
 
 import 'react-toastify/dist/ReactToastify.css'
 
@@ -34,7 +34,7 @@ export const UserRegistration = () => {
 
   const signup =
     (formData: any, history: any, errorM?: any) => async (dispatch: any) => {
-      let data;
+      let data
       try {
         // sign up the user
         data = await api.signUp(userData)
